pyproject.toml
[project]
name = "submission"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kagglehub[pandas-datasets]>=0.4.3",
    "pandas>=2.0.0,<3",
    "plotly>=5.18.0",
    "schedule>=1.2.2",
    "streamlit>=1.40.0",
    "psycopg[binary]>=3.2.6",
    "python-dotenv>=1.0.1",
    "langchain-groq>=1.1.2",
    "altair<5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
etl = "src.etl:main"
scheduler = "src.scheduler:start_scheduler"
init-db = "src.etl:init_database_cli"
agent-cli = "src.agent:main"
